Remove ability to create empty Grammars
This really shouldn't be possible. An empty Grammar is meaningless.
